DF-6084 fix all iOS warnings
Former-commit-id: 09fc16a4c3a2f87cf247e0227aa078f88082bc4e

+++ Sources/Internal/FileSystem/Logger.cpp
@@ -260,8 +260,8 @@ void Logger::FileLog(eLogLevel ll, const char8* text)
         {
             std::array<char8, 128> prefix;
             snprintf(&prefix[0], prefix.size(), "[%s] ", GetLogLevelString(ll));
-            file->Write(prefix.data(), strlen(prefix.data()));
-            file->Write(text, strlen(text));
+            file->Write(prefix.data(), static_cast<uint32>(strlen(prefix.data())));
+            file->Write(text, static_cast<uint32>(strlen(text)));
             file->Release();
         }
     }
