+++ Sources/Internal/Render/2D/Font.cpp
@@ -87,7 +87,7 @@ bool Font::IsEqual(const Font *font) const
 uint32 Font::GetHashCode()
 {
 	String rawHashString = GetRawHashString();
-	return CRC32::ForBuffer(rawHashString.c_str(), rawHashString.length());
+	return CRC32::ForBuffer(rawHashString.c_str(), static_cast<uint32>(rawHashString.length()));
 }
 
 String Font::GetRawHashString()
