+++ Sources/Internal/Render/Image/LibPVRHelper.cpp
@@ -1978,7 +1978,7 @@ bool LibPVRHelper::GetCRCFromMetaData(const PVRFile *pvrFile, uint32* outputCRC)
         
     bool crcRead = false;
         
-    uint32 metaDataCount = pvrFile->metaDatablocks.size();
+    uint32 metaDataCount = static_cast<uint32>(pvrFile->metaDatablocks.size());
     for(uint32 i = 0; i < metaDataCount; ++i)
     {
         const MetaDataBlock * block = pvrFile->metaDatablocks[i];
@@ -2220,7 +2220,7 @@ bool LibPVRHelper::WriteFileFromMipMapFiles(const FilePath & outputFilePath, con
 {
     DVASSERT(imgPaths.size());
         
-    int32 levelsCount = imgPaths.size();
+    int32 levelsCount = static_cast<int32>(imgPaths.size());
         
     Vector<PVRFile *> pvrFiles;
     pvrFiles.reserve(levelsCount);
@@ -2243,7 +2243,7 @@ bool LibPVRHelper::WriteFileFromMipMapFiles(const FilePath & outputFilePath, con
         
     uint8 * dataPtr = allCompressedData;
         
-    int32 pvrFilesCount = pvrFiles.size();
+    int32 pvrFilesCount = static_cast<int32>(pvrFiles.size());
     for(int32 i = 0; i < pvrFilesCount; ++i)
     {
         PVRFile * leveli = pvrFiles[i];
